+++ src/core/plan/matching/Greater_condition.cpp
@@ -2,7 +2,7 @@
 
 using namespace db07;
 
-bool Greater_condition::interpret(int compare_result) const
+bool Greater_condition::interpret(double compare_result) const
 {
 	return compare_result > 0;
 }
