+++ src/core/plan/matching/Less_condition.cpp
@@ -2,7 +2,7 @@
 
 using namespace db07;
 
-bool Less_condition::interpret(int compare_result) const
+bool Less_condition::interpret(double compare_result) const
 {
     return compare_result < 0;
 }
