+++ src/core/plan/matching/Less_equals_condition.cpp
@@ -1,6 +1,6 @@
 #include "Less_equals_condition.h"
 
-bool db07::Less_equals_condition::interpret(int compare_result) const
+bool db07::Less_equals_condition::interpret(double compare_result) const
 {
 	return compare_result <= 0;
 }
