+++ src/core/plan/matching/Unequals_condition.cpp
@@ -2,7 +2,7 @@
 
 using namespace db07;
 
-bool Unequals_condition::interpret(int compare_result) const
+bool Unequals_condition::interpret(double compare_result) const
 {
 	return compare_result != 0;
 }
