+++ src/core/plan/matching/Unequals_condition.h
@@ -6,7 +6,7 @@
 namespace db07 {
     class Unequals_condition : public Comparing_condition {
     protected:
-        bool interpret(int compare_result) const override;
+        bool interpret(double compare_result) const override;
 
     public:
         Unequals_condition(unsigned int column, std::unique_ptr<Value> value) :
